update src to have client folder so we can serve a backend to
