+++ src/client/components/common/header.css
@@ -1,0 +1,4 @@
+header {
+  height: 80px;
+  width: 100%;
+}
