+++ src/client/components/common/header.js
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import './header.css';
 // import child components
 import Navigation from './navigation';
