+++ src/client/index.js
@@ -3,8 +3,9 @@ import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 
 // import routes and store for use in render
-import routes from './routes';
+import Routes from './routes';
 import store from './store';
+// import NpmCheck from './components/npm-check';
 
 // this line throws an eslint error as unused var but we need to import otherwise
 // the indexeddb is never instantiated... TODO find a better way to manage this
@@ -22,7 +23,7 @@ import AppContainer from './containers/app-container';
     render(
       <Provider store={store}>
         <AppContainer>
-          <routes/>
+          <Routes />
         </AppContainer>
       </Provider>,
       document.getElementById('app')
