+++ src/client/reducers/app.js
