+++ src/client/utils/browser-history.js
