+++ test/unit/client/components/common/header.spec.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import TestUtils from 'react-addons-test-utils'
-import Navigation from '../../../../../src/components/common/navigation';
+import Header from '../../../../../src/client/components/common/header'
 
 function setup() {
 
   let renderer = TestUtils.createRenderer()
-  renderer.render(<Navigation/>)
+  renderer.render(<Header/>)
   const output = renderer.getRenderOutput()
 
   return {
@@ -16,12 +16,12 @@ function setup() {
 
 describe('components', () => {
 
-  describe('navigation', () => {
+  describe('headers', () => {
 
     it('should render correctly', () => {
       const { output } = setup();
 
-      expect(output.type).to.equal('nav');
+      expect(output.type).to.equal('header');
     });
 
   });
