+++ webpack.config.js
@@ -1,41 +1,15 @@
 var path = require('path');
 var webpack = require('webpack');
 var ExtractTextPlugin = require('extract-text-webpack-plugin');
-var fs = require('fs');
+var inject = require('./tasks/webpack/inject');
 var hotPort = 8000;
 
 var isDev = process.env.NODE_ENV !== 'production';
 
-function inject () {
-  this.plugin("done", function (stats) {
-    var replaceInFile = function (filePath, toReplace, replacement) {
-      var replacer = function (match) {
-        console.log('Replacing in %s: %s => %s', filePath, match, replacement);
-        return replacement;
-      };
-      var str = fs.readFileSync(filePath, 'utf8');
-      var out = str.replace(new RegExp(toReplace, 'g'), replacer);
-      fs.writeFileSync(filePath, out);
-    };
-
-    var hash = stats.hash; // Build's hash, found in `stats` since build lifecycle is done.
-
-    replaceInFile(path.join(path.join(__dirname, 'dist'), 'index.html'),
-        'bundle.js',
-        'bundle.' + hash + '.js'
-    );
-
-    replaceInFile(path.join(path.join(__dirname, 'dist'), 'index.html'),
-        'bundle.css',
-        'bundle.' + hash + '.css'
-    );
-  });
-}
-
 module.exports = {
   devtool: (isDev) ? 'eval' : 'source-map',
   entry: [
-    path.join(__dirname, 'src/index.js'),
+    path.join(__dirname, 'src/client/index.js'),
   ].concat(isDev ? [
     'webpack-dev-server/client?http://localhost:' + hotPort,
     'webpack/hot/dev-server'
